feat: removed 'CPLoad' command
This is because the 'colorscheme' command is linked to the colorscheme
event, which means that using CPLoad would have conflicted/caused errors
in plugins/users that were using this event. Furthermore, it made little
to no sense to use a custom command to load colorschemes for Catppuccino
when there is already a built-in one.