Fix lack of foreground text for `@text.todo` (#358)
`@text.todo` was lacking a foreground color so it would take the colour
from the comment - this in the latte theme lead to a pretty nasty
grey-on-orange theme. Compared to the other types like `@text.danger`,
it seems we're missing the `fg` property which this PR sets.

As a drive-by change, I also added a missing brace in the example code
for overrides in `README.md`.