fix(neotree): set correct color for inactive neotree window
NeoTree links its `NeoTreeNormalNc` highlight group to `NormalNc`. This
means that when the NeoTree window becomes inactive, it reuses the same
color that is set for `NormalNc`. This causes issues, as this will cause
the window to switch to the same color as the other windows, if dim is
not enabled.

To fix this, the `NeoTreeNormalNc` color is now explicitly set to the
same value as `NeoTreeNormal`.

fixes #476