fix: Airline warning section in Replace mode
The colors for the error and warning section of Airline were not applied
to replace mode. This commit fixes this issue by instead defining
the replace mode colors as a variation of the insert mode colors, where
these colors are already deifned. Note that the red color is only
applied to Airline sections A & B, but not on the right side, since the
red color is already used there for the error section. This is different
from Lualine, since there the error & warning sections don't exist.