feat: Add Terminal+Command mode colors to Airline
This adds support for different colors for Airline in terminal mode and
command mode, where before simply the normal mode colors where used.
For terminal mode, its simply a copy of insert mode. This makes sense
since it is basically insert mode just for the terminal buffer. In
Lualine, this is done the same way.
For command mode, it is a variation of normal mode. The orange color is
also used in Lualine for this purpose. Only the sections A & B (on the
left) are modified with this color, different from Lualine. Reason is
the same as for Replace mode; the orange color is already used on the
right for the warning section.