Refresh airline colors (#691)
* fix: Background color Airline section B.

Airline section B is used to show information regarding vcs, i.e. git
branch. This section was formatted the same as section C, so they were
indistinguishable.

This commit changes the bachground color to the same one used for
Lualine.

* fix: Airline warning section in Replace mode

The colors for the error and warning section of Airline were not applied
to replace mode. This commit fixes this issue by instead defining
the replace mode colors as a variation of the insert mode colors, where
these colors are already deifned. Note that the red color is only
applied to Airline sections A & B, but not on the right side, since the
red color is already used there for the error section. This is different
from Lualine, since there the error & warning sections don't exist.

* feat: Add Terminal+Command mode colors to Airline

This adds support for different colors for Airline in terminal mode and
command mode, where before simply the normal mode colors where used.
For terminal mode, its simply a copy of insert mode. This makes sense
since it is basically insert mode just for the terminal buffer. In
Lualine, this is done the same way.
For command mode, it is a variation of normal mode. The orange color is
also used in Lualine for this purpose. Only the sections A & B (on the
left) are modified with this color, different from Lualine. Reason is
the same as for Replace mode; the orange color is already used on the
right for the warning section.