feat(native_lsp): support new "Ok" diagnostics
Defaults to using the green color.

This was not added to the "LspDiagnostic*" highlights since I could not
find any documentation on whether the "Ok" diagnostic is implemented for
those. Indeed, I could not find any documentation whatsoever on the
LspDiagnostic* highlights (grepping the Neovim source turned up no
mentions of them either).