docs(README): Replace deprecated packer.nvim with built-in package manager
Motivation behind that is simple: I am migrating from lazy to built-in plugin manager. /jk

Basically packer.nvim is deprecated where the message in readme states: "NOTICE:

This repository is currently unmaintained. For the time being (as of August, 2023), it is recommended to use one of the following plugin managers instead: ...".

It's already deprecated for 2 years. That's why I thought it would be a good idea to replace packer.nvim with a new built-in plugin manager. One removed, one added.